Format 24 hour times as Times in Excel.
If you provide a String of the exact 24-hour format of "hh:mm:ss", this will format it so Excel recognizes it as a Time instead of just a String.

By default, it sets the format to the normal 24-hour format, "hh:mm:ss", but within Excel you could easily adjust this to show AM/PM, etc. When it was just a String you can't do that.

This is accomplished by:

A) Using a RegEx to determine if the String is in the "hh:mm:ss" format.

B) Determine the fraction of the day that the time represents, which is how Excel recognizes and formats Time values.