Add limit detection for 65,530 URLs in a worksheet and fallback to String if exceeded.
There are hard limits to the number of URLs you can have in a single Worksheet. You can find more on the Microsoft docs here:

https://support.office.com/en-us/article/excel-specifications-and-limits-1672b34d-7043-467e-8e27-269d656771c3?redirectSourcePath=%252fen-us%252farticle%252fExcel-specifications-and-limits-ca36e2dc-1f09-4620-b726-67c00b05040f&ui=en-US&rs=en-US&ad=US#ID0EBABAAA=Office_2007

But note this document says 66,530 and in practice I have seen a limit of 65,530. Maybe a typo or maybe something else going on. But I have decided to set the limit to 65,530 to fit my needs and be conservative.