Handle percent values.
If a string is provided that is a number (with or without decimal places) and ends in a percent sign ("%"), it will be converted to a decimal and formatted as a percentage in Excel.

The number of decimal places to use in Excel will be based on the number of decimal places provided in the String, defaulting to 2 if the source string does not have any.